+++ CHANGELOG.md
@@ -10,12 +10,18 @@ All notable changes, such as backward incompatibilities, will be documented in t
 
 ### Added
 
-- [A patch](./utils/speed_pqcsigs_in_default_provider.patch) to fix [this error](https://github.com/openssl/openssl/issues/27108).
+- [A patch](./utils/speed_pqcsigs_in_default_provider.patch) to fix [this error](https://github.com/openssl/openssl/issues/27108)
+  - and `speed_pqcsigs_in_default_provider_3_5_0.patch` for OpenSSL 3.5.0.
 - How to troubleshoot `TimeoutError: Navigation timeout of` in README.
+- `-r run_mode` option to [plot_openssl_speed_all.sh](./plot_openssl_speed_all.sh).
 
 ### Changed
 
 - GIT_CLONE="${GIT} clone"
+
+### Fixed
+
+- "not enough arguments to satisfy format string" error in `extract_rsa_sig ()`.
 
 ## [1.2.0]